Replace `post-capture-details-page` with `details-page`.
