--- src/app/features/home/post-capture-tab/post-capture-details/post-capture-details-routing.module.ts
@@ -1,16 +1,0 @@
-import { NgModule } from '@angular/core';
-import { RouterModule, Routes } from '@angular/router';
-import { PostCaptureDetailsPage } from './post-capture-details.page';
-
-const routes: Routes = [
-  {
-    path: '',
-    component: PostCaptureDetailsPage,
-  },
-];
-
-@NgModule({
-  imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule],
-})
-export class PostCaptureDetailsPageRoutingModule {}
