--- src/app/features/home/post-capture-tab/post-capture-details/post-capture-details.module.ts
@@ -1,10 +1,0 @@
-import { NgModule } from '@angular/core';
-import { SharedModule } from '../../../../shared/shared.module';
-import { PostCaptureDetailsPageRoutingModule } from './post-capture-details-routing.module';
-import { PostCaptureDetailsPage } from './post-capture-details.page';
-
-@NgModule({
-  imports: [SharedModule, PostCaptureDetailsPageRoutingModule],
-  declarations: [PostCaptureDetailsPage],
-})
-export class PostCaptureDetailsPageModule {}
