--- src/app/features/home/post-capture-tab/post-capture-details/post-capture-details.page.html
@@ -1,52 +1,0 @@
-<mat-toolbar color="primary" *transloco="let t">
-  <button (click)="back()" mat-icon-button>
-    <mat-icon>arrow_back</mat-icon>
-  </button>
-  <div class="toolbar-spacer"></div>
-  <button (click)="openContactSelectionDialog()" mat-icon-button>
-    <mat-icon>send</mat-icon>
-  </button>
-  <button (click)="openOptionsMenu()" mat-icon-button>
-    <mat-icon>more_vert</mat-icon>
-  </button>
-</mat-toolbar>
-
-<mat-list *ngrxLet="diaBackendAsset$ as diaBackendAsset">
-  <app-media
-    *ngrxLet="mediaSrc$ as mediaSrc"
-    [routerLink]="[
-      '/media-viewer',
-      mediaSrc,
-      { mimeType: diaBackendAsset.asset_file_mime_type }
-    ]"
-    [src]="mediaSrc"
-    [mimeType]="diaBackendAsset.asset_file_mime_type"
-  ></app-media>
-  <mat-list-item>
-    <mat-icon mat-list-icon>person</mat-icon>
-    <div mat-line>
-      {{
-        diaBackendAsset.source_transaction?.sender_name ||
-          diaBackendAsset.creator_name
-      }}
-    </div>
-  </mat-list-item>
-  <mat-list-item>
-    <mat-icon mat-list-icon>place</mat-icon>
-    <div mat-line (click)="openMap()">
-      {{ location$ | ngrxPush }}
-    </div>
-  </mat-list-item>
-  <mat-list-item>
-    <mat-icon mat-list-icon>access_time</mat-icon>
-    <div mat-line>
-      {{ diaBackendAsset.information.proof?.timestamp | date: 'long' }}
-    </div>
-  </mat-list-item>
-  <mat-list-item>
-    <mat-icon svgIcon="media-id-solid-black" mat-list-icon></mat-icon>
-    <div mat-line>
-      {{ diaBackendAsset.id }}
-    </div>
-  </mat-list-item>
-</mat-list>
