--- src/app/features/home/post-capture-tab/post-capture-details/post-capture-details.page.scss
@@ -1,21 +1,0 @@
-:host {
-  justify-content: start;
-}
-
-.toolbar-spacer {
-  flex: 1 1 auto;
-}
-
-mat-list {
-  overflow: scroll;
-  padding-top: 0;
-}
-
-.mat-list-base .mat-list-item.mat-list-item-with-avatar {
-  height: 40px;
-  padding: 0 20px;
-}
-
-video {
-  width: 100%;
-}
