--- src/app/features/home/post-capture-tab/post-capture-details/post-capture-details.page.spec.ts
@@ -1,25 +1,0 @@
-import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
-import { SharedTestingModule } from '../../../../shared/shared-testing.module';
-import { PostCaptureDetailsPage } from './post-capture-details.page';
-
-describe('PostCaptureDetailsPage', () => {
-  let component: PostCaptureDetailsPage;
-  let fixture: ComponentFixture<PostCaptureDetailsPage>;
-
-  beforeEach(
-    waitForAsync(() => {
-      TestBed.configureTestingModule({
-        declarations: [PostCaptureDetailsPage],
-        imports: [SharedTestingModule],
-      }).compileComponents();
-
-      fixture = TestBed.createComponent(PostCaptureDetailsPage);
-      component = fixture.componentInstance;
-      fixture.detectChanges();
-    })
-  );
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
